requirements.txt
pystray
Pillow
playwright
